+++ es-core/src/resources/Font.cpp
@@ -379,7 +379,8 @@ void Font::renderTextCache(TextCache* cache)
         auto vertexList = *it;
 
         Renderer::bindTexture(*it->textureIdPtr);
-        Renderer::drawTriangleStrips(&it->verts[0], it->verts.size());
+        Renderer::drawTriangleStrips(&it->verts[0],
+                static_cast<const unsigned int>(it->verts.size()));
     }
 }
